chore: rearrange files

+++ lab/00/code-1/src/main.rs
@@ -1,6 +1,6 @@
 use std::io;
 
-fn rin() -> i32 {
+fn rin() -> i64 {
     let mut input = String::new();
     io::stdin().read_line(&mut input).unwrap();
     input.trim().parse().unwrap()
